feat: 实现前端基础架构和核心功能模块
refactor(backend): 优化数据库查询和错误处理逻辑
fix(backend): 修复实例关系查询参数错误
perf(backend): 改进数据库连接池配置

feat(frontend): 添加模型管理、关系管理和SQL查询功能
style(frontend): 实现响应式布局和主题样式
docs: 更新.gitignore文件和Docker配置

fix: 修正跨域请求处理和权限验证中间件
perf: 优化模型字段类型校验和默认值处理

+++ backend/app/cmdb/models/modelFieldUnique.go
@@ -5,7 +5,7 @@ package models
 type ModelFieldUnique struct {
 	BaseModel
 	Fields      string `gorm:"column:fields; type:string; size:255; uniqueIndex:idx_fields_model_id; comment:字段" json:"fields" binding:"required"` // 唯一字段,联合唯一用英文逗号分隔                                                 // 字段分组ID
-	ModelId     int64  `gorm:"column:model_id; type:uint; uniqueIndex:idx_fields_model_id; comment:模型" json:"model_id" binding:"required"`         // 对应的模型ID
+	ModelId     uint   `gorm:"column:model_id; type:uint; uniqueIndex:idx_fields_model_id; comment:模型" json:"model_id" binding:"required"`         // 对应的模型ID
 	Description string `gorm:"column:description;type:text;comment:描述" json:"description"`                                                         // 描述
 }
